refactor(pi): split caveman extension modules
